requirements.txt
django
djangorestframework
django-stubs
pillow
django-debug-toolbar
pytils
mimesis
psycopg2
django-admin-geomap
geopy
django-tinymce
bleach